# pkg/metrics: Tune NewMetrics function