@mxinden 2018-10-29
Applied all comments @beorn7, thanks for the review.

Also looked into converting `WriteString` of length two with two `WriteByte` as mentioned [here](https://github.com/golang/go/issues/26264), but improvement seems not worth it.

@beorn7 @brancz any further comments?